move access permissions methods to security manager
