+++ superset/security.py
@@ -7,12 +7,16 @@ from __future__ import unicode_literals
 
 import logging
 
+from flask import g
 from flask_appbuilder.security.sqla import models as ab_models
+from flask_appbuilder.security.sqla.manager import SecurityManager
 from sqlalchemy import or_
 
-from superset import conf, db, sm
+from superset import appbuilder, conf, db, sm, sql_parse, utils
 from superset.connectors.connector_registry import ConnectorRegistry
+from superset.connectors.sqla.models import SqlaTable
 from superset.models import core as models
+
 
 READ_ONLY_MODEL_VIEWS = {
     'DatabaseAsync',
@@ -75,6 +79,136 @@ OBJECT_SPEC_PERMISSIONS = set([
     'datasource_access',
     'metric_access',
 ])
+
+
+class SupersetBaseDatasourceSecurityManager(SecurityManager):
+    """docstring for DatasourceSecurityManager"""
+
+    def can_access(self, permission_name, view_name, user=None):
+        if not user:
+            user = g.user
+        if user.is_anonymous():
+            return sm.is_item_public(permission_name, view_name)
+        return self._has_view_access(user, permission_name, view_name)
+
+    def all_datasource_access(self, user=None):
+        return self.can_access(
+            'all_datasource_access', 'all_datasource_access', user=user)
+
+    def database_access(self, database, user=None):
+        return (
+            self.can_access(
+                'all_database_access', 'all_database_access', user=user) or
+            self.can_access('database_access', database.perm, user=user)
+        )
+
+    def schema_access(self, datasource, user=None):
+        return (
+            self.database_access(datasource.database, user=user) or
+            self.all_datasource_access(user=user) or
+            self.can_access('schema_access', datasource.schema_perm, user=user)
+        )
+
+    def datasource_access(self, datasource, user=None):
+        return (
+            self.schema_access(datasource, user=user) or
+            self.can_access('datasource_access', datasource.perm, user=user)
+        )
+
+    def datasource_access_by_name(
+            self, database, datasource_name, schema=None):
+        if self.database_access(database) or self.all_datasource_access():
+            return True
+
+        schema_perm = utils.get_schema_perm(database, schema)
+        if schema and self.can_access('schema_access', schema_perm):
+            return True
+
+        datasources = ConnectorRegistry.query_datasources_by_name(
+            db.session, database, datasource_name, schema=schema)
+        for datasource in datasources:
+            if self.can_access('datasource_access', datasource.perm):
+                return True
+        return False
+
+    def datasource_access_by_fullname(
+            self, database, full_table_name, schema):
+        table_name_pieces = full_table_name.split('.')
+        if len(table_name_pieces) == 2:
+            table_schema = table_name_pieces[0]
+            table_name = table_name_pieces[1]
+        else:
+            table_schema = schema
+            table_name = table_name_pieces[0]
+        return self.datasource_access_by_name(
+            database, table_name, schema=table_schema)
+
+    def rejected_datasources(self, sql, database, schema):
+        superset_query = sql_parse.SupersetQuery(sql)
+        return [
+            t for t in superset_query.tables if not
+            self.datasource_access_by_fullname(database, t, schema)]
+
+    def user_datasource_perms(self):
+        datasource_perms = set()
+        for r in g.user.roles:
+            for perm in r.permissions:
+                if (
+                        perm.permission and
+                        'datasource_access' == perm.permission.name):
+                    datasource_perms.add(perm.view_menu.name)
+        return datasource_perms
+
+    def schemas_accessible_by_user(self, database, schemas):
+        if self.database_access(database) or self.all_datasource_access():
+            return schemas
+
+        subset = set()
+        for schema in schemas:
+            schema_perm = utils.get_schema_perm(database, schema)
+            if self.can_access('schema_access', schema_perm):
+                subset.add(schema)
+
+        perms = self.user_datasource_perms()
+        if perms:
+            tables = (
+                db.session.query(SqlaTable)
+                .filter(
+                    SqlaTable.perm.in_(perms),
+                    SqlaTable.database_id == database.id,
+                )
+                .all()
+            )
+            for t in tables:
+                if t.schema:
+                    subset.add(t.schema)
+        return sorted(list(subset))
+
+    def accessible_by_user(self, database, datasource_names, schema=None):
+        if self.database_access(database) or self.all_datasource_access():
+            return datasource_names
+
+        if schema:
+            schema_perm = utils.get_schema_perm(database, schema)
+            if self.can_access('schema_access', schema_perm):
+                return datasource_names
+
+        user_perms = self.user_datasource_perms()
+        user_datasources = ConnectorRegistry.query_datasources_by_permissions(
+            db.session, database, user_perms)
+        if schema:
+            names = {
+                d.table_name
+                for d in user_datasources if d.schema == schema}
+            return [d for d in datasource_names if d in names]
+        else:
+            full_names = {d.full_name for d in user_datasources}
+            return [d for d in datasource_names if d in full_names]
+
+
+dsm = conf.get('CUSTOM_DATASOURCE_SECURITY_MANAGER')
+if not dsm:
+    dsm = SupersetBaseDatasourceSecurityManager(appbuilder)
 
 
 def merge_perm(sm, permission_name, view_menu_name):
