+++ superset/views/base.py
@@ -20,9 +20,7 @@ from flask_babel import gettext as __
 from flask_babel import lazy_gettext as _
 import yaml
 
-from superset import appbuilder, conf, db, sm, sql_parse, utils
-from superset.connectors.connector_registry import ConnectorRegistry
-from superset.connectors.sqla.models import SqlaTable
+from superset import appbuilder, conf, sm, utils
 from superset.translations.utils import get_language_pack
 
 FRONTEND_CONF_KEYS = (
@@ -88,126 +86,6 @@ def get_user_roles():
 
 
 class BaseSupersetView(BaseView):
-    def can_access(self, permission_name, view_name, user=None):
-        if not user:
-            user = g.user
-        return utils.can_access(
-            appbuilder.sm, permission_name, view_name, user)
-
-    def all_datasource_access(self, user=None):
-        return self.can_access(
-            'all_datasource_access', 'all_datasource_access', user=user)
-
-    def database_access(self, database, user=None):
-        return (
-            self.can_access(
-                'all_database_access', 'all_database_access', user=user) or
-            self.can_access('database_access', database.perm, user=user)
-        )
-
-    def schema_access(self, datasource, user=None):
-        return (
-            self.database_access(datasource.database, user=user) or
-            self.all_datasource_access(user=user) or
-            self.can_access('schema_access', datasource.schema_perm, user=user)
-        )
-
-    def datasource_access(self, datasource, user=None):
-        return (
-            self.schema_access(datasource, user=user) or
-            self.can_access('datasource_access', datasource.perm, user=user)
-        )
-
-    def datasource_access_by_name(
-            self, database, datasource_name, schema=None):
-        if self.database_access(database) or self.all_datasource_access():
-            return True
-
-        schema_perm = utils.get_schema_perm(database, schema)
-        if schema and self.can_access('schema_access', schema_perm):
-            return True
-
-        datasources = ConnectorRegistry.query_datasources_by_name(
-            db.session, database, datasource_name, schema=schema)
-        for datasource in datasources:
-            if self.can_access('datasource_access', datasource.perm):
-                return True
-        return False
-
-    def datasource_access_by_fullname(
-            self, database, full_table_name, schema):
-        table_name_pieces = full_table_name.split('.')
-        if len(table_name_pieces) == 2:
-            table_schema = table_name_pieces[0]
-            table_name = table_name_pieces[1]
-        else:
-            table_schema = schema
-            table_name = table_name_pieces[0]
-        return self.datasource_access_by_name(
-            database, table_name, schema=table_schema)
-
-    def rejected_datasources(self, sql, database, schema):
-        superset_query = sql_parse.SupersetQuery(sql)
-        return [
-            t for t in superset_query.tables if not
-            self.datasource_access_by_fullname(database, t, schema)]
-
-    def user_datasource_perms(self):
-        datasource_perms = set()
-        for r in g.user.roles:
-            for perm in r.permissions:
-                if (
-                        perm.permission and
-                        'datasource_access' == perm.permission.name):
-                    datasource_perms.add(perm.view_menu.name)
-        return datasource_perms
-
-    def schemas_accessible_by_user(self, database, schemas):
-        if self.database_access(database) or self.all_datasource_access():
-            return schemas
-
-        subset = set()
-        for schema in schemas:
-            schema_perm = utils.get_schema_perm(database, schema)
-            if self.can_access('schema_access', schema_perm):
-                subset.add(schema)
-
-        perms = self.user_datasource_perms()
-        if perms:
-            tables = (
-                db.session.query(SqlaTable)
-                .filter(
-                    SqlaTable.perm.in_(perms),
-                    SqlaTable.database_id == database.id,
-                )
-                .all()
-            )
-            for t in tables:
-                if t.schema:
-                    subset.add(t.schema)
-        return sorted(list(subset))
-
-    def accessible_by_user(self, database, datasource_names, schema=None):
-        if self.database_access(database) or self.all_datasource_access():
-            return datasource_names
-
-        if schema:
-            schema_perm = utils.get_schema_perm(database, schema)
-            if self.can_access('schema_access', schema_perm):
-                return datasource_names
-
-        user_perms = self.user_datasource_perms()
-        user_datasources = ConnectorRegistry.query_datasources_by_permissions(
-            db.session, database, user_perms)
-        if schema:
-            names = {
-                d.table_name
-                for d in user_datasources if d.schema == schema}
-            return [d for d in datasource_names if d in names]
-        else:
-            full_names = {d.full_name for d in user_datasources}
-            return [d for d in datasource_names if d in full_names]
-
     def common_bootsrap_payload(self):
         """Common data always sent to the client"""
         messages = get_flashed_messages(with_categories=True)
@@ -334,6 +212,7 @@ class SupersetFilter(BaseFilter):
     """
 
     def get_user_roles(self):
+        # can this be cached?
         return get_user_roles()
 
     def get_all_permissions(self):
